Convert to Python3

+++ pydevd_attach_to_process/winappdbg/win32/context_i386.py
@@ -55,12 +55,12 @@ EXCEPTION_EXECUTE_FAULT     = 8     # exception caused by an instruction fetch
 CONTEXT_i386                = 0x00010000    # this assumes that i386 and
 CONTEXT_i486                = 0x00010000    # i486 have identical context records
 
-CONTEXT_CONTROL             = (CONTEXT_i386 | long(0x00000001)) # SS:SP, CS:IP, FLAGS, BP
-CONTEXT_INTEGER             = (CONTEXT_i386 | long(0x00000002)) # AX, BX, CX, DX, SI, DI
-CONTEXT_SEGMENTS            = (CONTEXT_i386 | long(0x00000004)) # DS, ES, FS, GS
-CONTEXT_FLOATING_POINT      = (CONTEXT_i386 | long(0x00000008)) # 387 state
-CONTEXT_DEBUG_REGISTERS     = (CONTEXT_i386 | long(0x00000010)) # DB 0-3,6,7
-CONTEXT_EXTENDED_REGISTERS  = (CONTEXT_i386 | long(0x00000020)) # cpu specific extensions
+CONTEXT_CONTROL             = (CONTEXT_i386 | int(0x00000001)) # SS:SP, CS:IP, FLAGS, BP
+CONTEXT_INTEGER             = (CONTEXT_i386 | int(0x00000002)) # AX, BX, CX, DX, SI, DI
+CONTEXT_SEGMENTS            = (CONTEXT_i386 | int(0x00000004)) # DS, ES, FS, GS
+CONTEXT_FLOATING_POINT      = (CONTEXT_i386 | int(0x00000008)) # 387 state
+CONTEXT_DEBUG_REGISTERS     = (CONTEXT_i386 | int(0x00000010)) # DB 0-3,6,7
+CONTEXT_EXTENDED_REGISTERS  = (CONTEXT_i386 | int(0x00000020)) # cpu specific extensions
 
 CONTEXT_FULL = (CONTEXT_CONTROL | CONTEXT_INTEGER | CONTEXT_SEGMENTS)
